+++ flaskr/auth.py
@@ -57,7 +57,7 @@ def login():
         if error is None:
             session.clear()
             session['user_id'] = user['id']
-            return redirect(url_for('messages.splash'))
+            return redirect(url_for("index"))
 
         flash(error)
 
@@ -77,7 +77,7 @@ def load_logged_in_user():
 @bp.route('/logout')
 def logout():
     session.clear()
-    return redirect(url_for('messages.index'))
+    return redirect(url_for("index"))
 
 def login_required(view):
     @functools.wraps(view)
