+++ templates/base.html
@@ -1,12 +1,12 @@
 <!doctype html>
-<title>{% block title %}{% endblock %} - Secret</title>
+<title>{% block title %}{% endblock %}</title>
 <link rel="stylesheet" href="{{ url_for('static', filename='style.css') }}">
 <nav>
-  <h1>Secret</h1>
+  <h1><a href="{{ url_for('steno.index') }}">Secret</a></h1>
   <ul>
     {% if g.user %}
       <li><span class="username">{{ g.user['username'] }}</span>
-      <li><a href="{{ url_for('messages.messages') }}">Messages</a>
+<!--      <li><a href="{{ url_for('steno.messages') }}">Messages</a>-->
       <li><a href="{{ url_for('auth.logout') }}">Log Out</a>
     {% else %}
       <li><a href="{{ url_for('auth.register') }}">Register</a>
