--- templates/messages/index.html
@@ -1,5 +1,0 @@
-{% extends 'base.html' %}
-
-{% block header %}
-  <h1>{% block title %}Log in for access.{% endblock %}</h1>
-{% endblock %}
