+++ templates/steno/create.html
@@ -5,7 +5,7 @@
 {% endblock %}
 
 {% block content %}
-  <form method="messages">
+  <form method="steno">
     <label for="title">Title</label>
     <input name="title" id="title" value="{{ request.form['title'] }}" required>
     <label for="body">Body</label>
