+++ templates/steno/messages.html
@@ -3,23 +3,23 @@
 {% block header %}
   <h1>{% block title %}Messages{% endblock %}</h1>
   {% if g.user %}
-    <a class="action" href="{{ url_for('messages.create') }}">New</a>
+    <a class="action" href="{{ url_for('steno.create') }}">New</a>
   {% endif %}
 {% endblock %}
 
 {% block content %}
-  {% for messages in messages %}
-    <article class="messages">
+  {% for steno in steno %}
+    <article class="steno">
       <header>
         <div>
-          <h1>{{ messages['title'] }}</h1>
-          <div class="about">by {{ messages['username'] }} on {{ messages['created'].strftime('%Y-%m-%d') }}</div>
+          <h1>{{ steno['title'] }}</h1>
+          <div class="about">by {{ steno['username'] }} on {{ steno['created'].strftime('%Y-%m-%d') }}</div>
         </div>
-<!--        {% if g.user['id'] == messages['owner_id'] %}
+<!--        {% if g.user['id'] == steno['owner_id'] %}
           <a class="action" href="{{ url_for('blog.update', id=post['id']) }}">Edit</a>
         {% endif %}-->
       </header>
-      <p class="body">{{ messages['body'] }}</p>
+      <p class="body">{{ steno['body'] }}</p>
     </article>
     {% if not loop.last %}
       <hr>
